pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cfa-factory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.4.0",
    "google-adk>=1.21.0",
    "google-genai>=1.56.0",
    "loguru>=0.7.3",
    "numpy>=2.4.0",
    "openai>=2.14.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "rich>=14.2.0",
    "typer>=0.21.0",
]

[project.scripts]
cfa = "cfa_factory.cli.main:main"
